requirements.txt
# ============================================================
# AdamoPay Partner Manager — Dependencias
# ============================================================

# Framework web
streamlit>=1.32.0

# Base de datos
SQLAlchemy>=2.0.0          # ORM / query builder (compatible SQLite + PostgreSQL)
alembic>=1.13.0            # Migraciones de esquema

# Validación de datos
pydantic>=2.6.0
email-validator>=2.1.0

# Seguridad (hashing de contraseñas)
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0   # JWT tokens

# Variables de entorno
python-dotenv>=1.0.0

# Utilidades de fechas
python-dateutil>=2.9.0

# Visualización (gráficas del dashboard)
plotly>=5.20.0
pandas>=2.2.0

# Exportación a Excel (reportes)
openpyxl>=3.1.0
xlsxwriter>=3.2.0

# Tests
pytest>=8.0.0
pytest-cov>=5.0.0

# PostgreSQL (producción — Railway)
psycopg2-binary>=2.9.9
